# lib/db, lib/model: Keep need stats in metadata (ref #5899)