--- a/lib/db/schemaupdater.go
+++ b/lib/db/schemaupdater.go
@@ -22,10 +22,10 @@ import (
 //   6: v0.14.50
 //   7: v0.14.53
 //   8-9: v1.4.0
-//   10: v1.4.1
+//   10: v1.5.0
 const (
 	dbVersion             = 10
-	dbMinSyncthingVersion = "v1.4.1"
+	dbMinSyncthingVersion = "v1.5.0"
 )
 
 type databaseDowngradeError struct {