--- a/lib/db/structs.proto
+++ b/lib/db/structs.proto
@@ -16,6 +16,7 @@ message FileVersion {
     protocol.Vector version = 1 [(gogoproto.nullable) = false];
     bytes           device  = 2;
     bool            invalid = 3;
+    bool            deleted = 4;
 }
 
 message VersionList {