This is the "main thing" from which I separated #6411 and #6412 (still in this PR as the first two commits). I also kept the rest as three commits for now, as they address separate issues and thus makes reviewing potentially easier - they should still be merged together in the end.

### Purpose

The purpose of this PR is to get rid of the frequent db access due to folder summaries/API completion queries. These require the global list and unmarshalling lots of stuff. I added a new metadata flag for needed files. Thus when a file is added to the db, that flag is updated and the summary/db query can just directly access those values from memory. Only if the actual list of needed files is required (i.e. user opens the list in UI) the db needs to be accessed.

### Testing

I extended tests in #6402 to cover more ground on "needs". And I now extended all these tests to look at both the metadata and db when checking needs.

I have also done some manual testing a while ago and before merging want to run this on my main setup.