@calmh 2020-03-16
Does this transform every file `put` into n `get`s, one for each device in the setup? So if I have a thousand devices, every update becomes a thousand reads plus, potentially, a thousand puts of updated metadata?