@imsodin 2020-03-16
Database access stays the same, max two: The new and old global file (and usually only one, as the inserted one is already the global one in most cases). What changes is that now the `VersionList`, that is anyway loaded, gets checked for every connected device and the metadata updated. So when the global version gets higher, there will be need checks and metadata bumps for every device, but that's in-memory only.

But good point, need to check benchmarks, didn't do that yet.