@imsodin 2020-03-29
I added a benchmark for updating with lots of remotes: There's 10 additional remotes with half of them on the old version, the other half on the new version. Then I update another device (local or a remote) between those two sets of files (as in existing benchs). I don't see any relevant change running it for 10s. While I expected no significant change (as db operations should be much costlier than a single loop over all devices with a few comparisons and writes to memory), they seem suspiciously close:

master:
```
BenchmarkUpdate100Changed10Remotes
BenchmarkUpdate100Changed10Remotes-4         	14577836	      1601 ns/op	     448 B/op	      16 allocs/op
BenchmarkUpdate100ChangedRemote10Remotes
BenchmarkUpdate100ChangedRemote10Remotes-4   	14450607	      1723 ns/op	     448 B/op	      15 allocs/op
```

PR:
```
BenchmarkUpdate100Changed10Remotes
BenchmarkUpdate100Changed10Remotes-4         	14166939	      1595 ns/op	     448 B/op	      16 allocs/op
BenchmarkUpdate100ChangedRemote10Remotes
BenchmarkUpdate100ChangedRemote10Remotes-4   	15173292	      1593 ns/op	     448 B/op	      15 allocs/op
```

A benchmark for getting need stats before and now feels pretty pointless: Now it's looking up a struct in memory, before iterating/unmarshalling the entire global prefix of the db doing lots of checks.