@calmh 2020-03-29
> A benchmark for getting need stats before and now feels pretty pointless

but that's the one that shows the awesomeness 😿