@imsodin 2020-03-29
> but that's the one that shows the awesomeness crying_cat_face

Brace for the awesomeness:

master:

```
BenchmarkUpdate100Changed10Remotes
BenchmarkUpdate100Changed10Remotes-4         	 7592064	      1578 ns/op	     448 B/op	      16 allocs/op
BenchmarkUpdate100ChangedRemote10Remotes
BenchmarkUpdate100ChangedRemote10Remotes-4   	 6654831	      2070 ns/op	     448 B/op	      15 allocs/op
BenchmarkNeedCount
BenchmarkNeedCount-4                         	    2136	   5576769 ns/op	  905676 B/op	   14089 allocs/op
```

PR:

```
BenchmarkUpdate100Changed10Remotes
BenchmarkUpdate100Changed10Remotes-4         	 7089142	      1775 ns/op	     448 B/op	      16 allocs/op
BenchmarkUpdate100ChangedRemote10Remotes
BenchmarkUpdate100ChangedRemote10Remotes-4   	 6603822	      1782 ns/op	     448 B/op	      15 allocs/op
BenchmarkNeedCount
BenchmarkNeedCount-4                         	255457330	        47.7 ns/op	       0 B/op	       0 allocs/op
```