@imsodin 2020-04-16
Most of the global update stuff is optimisations. I thought about in a first step doing a version that just stupidly checks everything and doesn't take any shortcuts - I can look into that again.