@imsodin 2020-04-30
I could make smaller, functional but on their own useless PRs against a feature branch to make it shorter evenings with smaller creates (or equally sized ones and funnier evenings). Essentially the first few commits here, plus hopefully i can split the last of the first ones up further.