@calmh 2020-05-11
Rerunning with `STRECHECKDBEVERY=1s` to force metadata recalc brought it back in line...