@imsodin 2020-05-11
The migration does not trigger a db check/meta recalc itself, but requires at least a meta recalc. Which on upgrade will happen, as schema upgrades come before the version comparison, leading to the check and recalc part. However that should only cause incorrect metadata, not out of sync - looking into that.