@imsodin 2020-05-11
That's it, could reproduce before (d1b13b7), is working correctly now:  
Running a dev build, the check/recalc on upgrade did not happen. This check recalculates the amount of needed files, which should be zero, but isn't due to the migration. When you click on the out of sync items list, it is empty, because the actually stored list of files that are out of sync is correct. Then recalculating metadata the two get in sync again.  
Now the migration triggers a meta recalc regardless of upgrade or not, by deleting the metadata in db.