@AudriusButkevicius 2020-05-11
Is v1.6.0 rc going out tomorrow, given sudden influx of reviews?